changes possible
Connect one square wave to an interrupt-capable pin on the Arduino. Let's call this pin "Pin1". This will trigger an interrupt whenever the rising edge of the wave occurs.

Connect the second square wave to another interrupt-capable pin on the Arduino. Let's call this pin "Pin2". This will also trigger an interrupt whenever the rising edge of the second wave occurs.

In your Arduino sketch, set up the interrupt service routines (ISRs) for both Pin1 and Pin2. Inside each ISR, you can capture the current value of the Arduino's timer.

Calculate the time difference between the two rising edges by subtracting the timer value captured in the ISR for Pin1 from the timer value captured in the ISR for Pin2. This will give you the time elapsed between the rising edges of the two square waves.
In this example, replace "Pin1" and "Pin2" with the actual pin numbers you are using to connect the square waves to the Arduino. The setup() function initializes the pins and attaches the interrupt service routines. The loop() function checks if both rising edges have been detected, calculates the time difference, and performs any necessary processing. The handleWave1Interrupt() and handleWave2Interrupt() functions are the interrupt service routines that capture the timer values when the rising edges occur.

Remember to connect the ground of the square wave sources with the Arduino's ground to have a common reference. Also, make sure to initialize the serial communication in your setup() function (Serial.begin(baudRate)) if you want to output the results to the serial monitor.

Keep in mind that this approach assumes the square waves have a fixed frequency and a consistent phase relationship. If the frequency or phase varies, additional measures might be required to handle these variations.